Fix spelling mistakes
config.yaml instead of ./config.yaml, because it doesn't work. A script's language should be constant, or "should never change" as to why, its just a waste of time, typescript is good enough for anything here, the only valid reason i know of to change language, is to use features that are not available nor recreatable. Every time I changed "have" to "had", I was not sure that you have done that yet. A LOT of times you use "many" are in the wrong place, you should only use it for "how many", use instead "a lot". "needed" some times it just doesn't work, instead use "required". Added foo@bar.com as an email example. Get grammarly to help you out.